[Feature & Refactoring] Add Todo Success Message - Add todo success status - Add Success message when deleting todo - Refactoring for message handling and Change TodoSnackbar componenet name - Delete uuid dependency
