+++ package.json
@@ -40,8 +40,7 @@
     "react-spinners": "^0.10.6",
     "react-tooltip": "^4.2.13",
     "react-use": "^17.1.1",
-    "recoil": "^0.1.2",
-    "uuid": "^8.3.2"
+    "recoil": "^0.1.2"
   },
   "devDependencies": {
     "@babel/core": "^7.12.10",
