+++ src/hooks/useRemoveCallback.js
@@ -3,7 +3,8 @@ import { useRecoilCallback } from 'recoil';
 import isLoadingAtom from '../recoil/common/atom';
 import { todosResultAtom, todoWithRemove } from '../recoil/todos';
 
-import { todoErrorMessage } from '../utils/utils';
+import { TODO_SUCCESS } from '../utils/constants/messages';
+import { todoErrorMessage } from '../utils/errorMessageHandling';
 
 const useRemoveCallback = () => useRecoilCallback(({ snapshot, set, reset }) => async (id) => {
   set(isLoadingAtom, true);
@@ -16,6 +17,7 @@ const useRemoveCallback = () => useRecoilCallback(({ snapshot, set, reset }) => 
       (prevState) => ({
         ...prevState,
         todos: prevState.todos.filter((todo) => todo._id !== id),
+        todoSuccess: TODO_SUCCESS.DELETE,
       }),
     );
   } catch (error) {
